rename init value mixin
